refactor!: libvcs.utils -> libvcs._internal
