+++ libvcs/_internal/dataclasses.py
