+++ libvcs/cmd/hg.py
@@ -3,7 +3,7 @@ import pathlib
 from typing import Optional, Sequence, Union
 
 from ..types import StrOrBytesPath, StrPath
-from libvcs.utils.run import run
+from libvcs._internal.run import run
 
 _CMD = Union[StrOrBytesPath, Sequence[StrOrBytesPath]]
