+++ libvcs/cmd/svn.py
@@ -2,7 +2,7 @@ import pathlib
 from typing import Literal, Optional, Sequence, Union
 
 from ..types import StrOrBytesPath, StrPath
-from libvcs.utils.run import run
+from libvcs._internal.run import run
 
 _CMD = Union[StrOrBytesPath, Sequence[StrOrBytesPath]]
